requirements.txt
streamlit==1.35.0
langchain==0.2.0
langchain-community==0.2.0
langchain-groq==0.1.0
PyPDF2==3.0.1
docx2txt==0.8
faiss-cpu==1.8.0
sentence-transformers==2.7.0
pandas==2.2.2
plotly==5.22.0
python-dotenv==1.0.1
torch==2.3.0
requests==2.32.3
beautifulsoup4==4.12.3
PyJWT==2.8.0
chromadb==0.5.0